Modify formal parameters of the message processing function and move declaration location of session pool auxiliary functions

+++ include/backend_proto.h
@@ -34,30 +34,30 @@ int backend_proxy_msg_response(uint8_t *msg);
 int backend_proxy_dev_msg_process(uint8_t *msg);
 int backend_proxy_dev_msg_response(uint8_t *msg);
 
-int backend_proxy_dev_msg_process_ver1(uint32_t msg_type, uint32_t msg_id, uint32_t action_type, uint32_t payload_len, uint8_t *msg_payload);
-int backend_proxy_dev_msg_process_enable_ver1(uint32_t payload_len, uint8_t *msg_payload);
-int backend_proxy_dev_msg_process_disable_ver1(uint32_t payload_len, uint8_t *msg_payload);
-int backend_proxy_dev_msg_process_query_ver1(uint32_t payload_len, uint8_t *msg_payload);
+int backend_proxy_dev_msg_process_ver1(uint16_t msg_type, uint16_t msg_id, uint16_t action_type, uint16_t payload_len, uint8_t *msg_payload);
+int backend_proxy_dev_msg_process_enable_ver1(uint16_t payload_len, uint8_t *msg_payload);
+int backend_proxy_dev_msg_process_disable_ver1(uint16_t payload_len, uint8_t *msg_payload);
+int backend_proxy_dev_msg_process_query_ver1(uint16_t payload_len, uint8_t *msg_payload);
 
 
 int backend_proxy_strgy_msg_process(uint8_t *msg);
 int backend_proxy_strgy_msg_response(uint8_t *msg);
-int backend_proxy_strgy_msg_process_ver1(uint32_t msg_type, uint32_t msg_id, uint32_t action_type, uint32_t payload_len, uint8_t *msg_payload);
-int backend_proxy_strgy_msg_process_set_ver1(uint32_t payload_len, uint8_t *msg_payload);
-int backend_proxy_strgy_msg_process_query_ver1(uint32_t payload_len, uint8_t *msg_payload);
+int backend_proxy_strgy_msg_process_ver1(uint16_t msg_type, uint16_t msg_id, uint16_t action_type, uint16_t payload_len, uint8_t *msg_payload);
+int backend_proxy_strgy_msg_process_set_ver1(uint16_t payload_len, uint8_t *msg_payload);
+int backend_proxy_strgy_msg_process_query_ver1(uint16_t payload_len, uint8_t *msg_payload);
 
 
-int backend_proxy_sess_msg_process(uint32_t frontend_sess_id, uint32_t backend_sess_id, uint8_t *msg);
+int backend_proxy_sess_msg_process(uint16_t frontend_sess_id, uint16_t backend_sess_id, uint8_t *msg);
 int backend_proxy_sess_msg_response(uint8_t *msg);
-int backend_proxy_sess_msg_process_ver1(uint32_t frontend_sess_id, uint32_t backend_sess_id, uint32_t msg_type, 
-                                        uint32_t action_type, uint16_t ip_version, uint32_t payload_len, 
+int backend_proxy_sess_msg_process_ver1(uint16_t frontend_sess_id, uint16_t backend_sess_id, uint16_t msg_type, 
+                                        uint16_t action_type, uint16_t ip_version, uint16_t payload_len, 
                                         uint8_t *msg_payload);
-int backend_proxy_sess_msg_process_create_ver1(uint32_t frontend_sess_id, uint32_t backend_sess_id, uint16_t ip_version, uint32_t payload_len, uint8_t *msg_payload);
-int backend_proxy_sess_msg_process_close_ver1(uint32_t frontend_sess_id, uint32_t backend_sess_id, uint32_t payload_len, uint8_t *msg_payload);
+int backend_proxy_sess_msg_process_create_ver1(uint16_t frontend_sess_id, uint16_t backend_sess_id, uint16_t ip_version, uint16_t payload_len, uint8_t *msg_payload);
+int backend_proxy_sess_msg_process_close_ver1(uint16_t frontend_sess_id, uint16_t backend_sess_id, uint16_t payload_len, uint8_t *msg_payload);
 
 
 
-int backend_proxy_data_msg_prosess(uint32_t frontend_sess_id, uint32_t backend_sess_id, uint8_t *msg);
+int backend_proxy_data_msg_prosess(uint16_t frontend_sess_id, uint16_t backend_sess_id, uint8_t *msg);
 int backend_proxy_data_msg_recv(struct BackendSession *sess, uint8_t *msg);
 int backend_proxy_data_msg_send(struct BackendSession *sess, uint8_t *msg);
